Change folder structure
- Move source files to src/;
- Move tests to tests/;
- Add phpcs
