--- Tests/bootstrap.php
@@ -1,3 +1,0 @@
-<?php
-
-require __DIR__.'/../vendor/autoload.php';
