+++ composer.json
@@ -12,6 +12,11 @@
             "email": "ludo.fleury@gmail.com"
         }
     ],
+    "scripts": {
+        "phpunit": "vendor/bin/phpunit",
+        "phpcs": "vendor/bin/phpcs --standard=phpcs.xml.dist src/ tests/",
+        "phpcs-fix": "vendor/bin/phpcbf --standard=phpcs.xml.dist src/ tests/"
+    },
     "require": {
         "php": ">=5.6",
         "guzzle/guzzle": "3.*",
@@ -23,11 +28,17 @@
         "symfony/yaml": "~2.1|~3.0|^4.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "~5.7"
+        "phpunit/phpunit": "~5.7",
+        "squizlabs/php_codesniffer": "^3.12"
     },
     "autoload": {
         "psr-4": {
-            "Playbloom\\Bundle\\GuzzleBundle\\": ""
+            "Playbloom\\Bundle\\GuzzleBundle\\": "src/"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Playbloom\\Bundle\\GuzzleBundle\\Tests\\": "tests/"
         }
     },
     "extra": {
