+++ src/DependencyInjection/Compiler/ClientPluginPass.php
