+++ src/Resources/public/font/fontawesome/fontawesome-webfont.eot
