+++ tests/bootstrap.php
@@ -1,0 +1,3 @@
+<?php
+
+require __DIR__ . '/../vendor/autoload.php';
